// frontend/src/core/api/generated/model/attendanceEntryView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { AttendanceEntryViewReason } from './attendanceEntryViewReason'
import type { AttendanceEntryViewStatus } from './attendanceEntryViewStatus'
import type { WorkerView } from './workerView'

export interface AttendanceEntryView {
  worker: WorkerView
  status: AttendanceEntryViewStatus
  /** @nullable */
  reason?: AttendanceEntryViewReason
  /** @nullable */
  note?: string | null
}